gitlab-ci: Push a `latest` tag to Docker repositories
When doing a `docker pull`, if a tag is not specified, it defaults to
`latest`. We currently don't push such a tag, so it can cause confusion
in some cases (e.g. https://github.com/albertito/chasquid/issues/21).

To help prevent this, make the docker automatic builds for `master` also
apply the `latest` tag.

As a part of this, update the obsolete `$CI_BUILD_REF_NAME` to
`$CI_COMMIT_REF_NAME`.