github: Use the latest Go version for govulncheck
govulncheck will complain when run on older Go versions that are known
to have security issues. Github setup-go action has a cache for the
"latest" go version and sometimes the latest isn't really used.

So sometimes, this generates false positives because govulncheck is run
on an older Go version, which is not a problem in chasquid itself.

To prevent those false positives, this patch asks the Github CI to
always check for the latest Go version when running govulncheck.