Update minimum supported Go version 1.15 -> 1.17
The latest version of some packages we depend on (golang.org/x/sys,
indirectly) do not support 1.15 anymore.

We kept 1.15 because it's the version on Debian stable, but that release
is more than 2y old, and newer versions are readily available in Debian
stable backports.

So this patch updates the minimum supported Go version to 1.17.