ci: Remove Cirrus CI configuration
Unfortunately, but understandably, Cirrus CI no longer offers enough
free compute credits to run our tests reliably.

They're currently only used to run the Go tests on FreeBSD.

In the future, this might be replaced with something else; but until a
proper replacement can be found and tested, remove it to avoid
introducing noise in the CI results.