ci: Fix docker push of `latest` tag
The `latest` tag is meant to track the `main` branch, but I just noticed
it hasn't been pushed out in a while. This is because the conditional
gating the push on the branch being `main` is incorrect.

This patch fixes the problem by using the correct conditional on the
branch name.